+++ components/src/lib/Modal/Modal.module.css
@@ -1,30 +1,3 @@
-.modalHeader {
-  display: flex;
-  justify-content: space-between;
-}
-
-.modalBody {
-  margin: 16px 0px;
-  flex-grow: 1;
-  overflow-y: auto;
-}
-
-.modalFooter {
-  display: flex;
-}
-
-.row {
-  flex-direction: row;
-  align-items: center;
-  column-gap: 8px;
-}
-
-.column {
-  flex-direction: column;
-  justify-content: center;
-  row-gap: 8px;
-}
-
 .modalLayout {
   position: fixed;
   top: 0;
@@ -35,16 +8,16 @@
   z-index: 999;
 }
 
+.modalLayout.top {
+  align-items: flex-start;
+}
+
 .modalLayout.center {
   align-items: center;
 }
 
 .modalLayout.bottom {
   align-items: flex-end;
-}
-
-.modalLayout.top {
-  align-items: flex-start;
 }
 
 .modalBackdrop {
