+++ components/src/lib/Modal/Modal.type.ts
@@ -1,6 +1,5 @@
 import type { Size } from '../types/common';
 
-export type Direction = 'row' | 'column';
 export type Position = 'center' | 'bottom' | 'top';
 export type BackDropType = 'transparent' | 'blur' | 'opaque';
 
@@ -12,8 +11,4 @@ export interface ModalProps {
   size?: Size;
   shadow?: boolean;
   animation?: boolean;
-}
-
-export interface ModalFooterProps {
-  direction?: Direction;
 }
