+++ components/src/lib/components/ModalLayout/ModalLayout.type.ts
@@ -1,0 +1,5 @@
+export type Direction = 'row' | 'column';
+
+export interface ModalFooterProps {
+  direction?: Direction;
+}
